# swaggerOutput.json
{
  "swagger": "2.0",
  "info": {
    "title": "API de Tarefas",
    "description": "API criada durante o curso de Sistemas de Informação",
    "version": "3.0.0"
  },
  "host": "localhost:3000",
  "basePath": "/",
  "schemes": [
    "http",
    "https"
  ],
  "paths": {
    "/tarefas": {
      "get": {
        "tags": [
          "Tarefas"
        ],
        "summary": "Requisite todas as tarefas.",
        "description": "Descrição mais detalhada da rota aqui!",
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "post": {
        "tags": [
          "Tarefas"
        ],
        "description": "",
        "parameters": [
          {
            "name": "tarefa",
            "in": "body",
            "description": "Dados enviados para cadastrar a tarefa:",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "nome": {
                  "type": "string",
                  "example": "Nome da Tarefa"
                },
                "descricao": {
                  "type": "string",
                  "example": "Descrição da tarefa"
                },
                "status": {
                  "type": "string",
                  "example": "Status"
                }
              },
              "required": [
                "nome",
                "descricao",
                "status"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          }
        }
      }
    },
    "/tarefas/{id}": {
      "get": {
        "tags": [
          "Tarefas"
        ],
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "delete": {
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          },
          "400": {
            "description": "Bad Request"
          }
        }
      },
      "put": {
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          }
        }
      },
      "patch": {
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          }
        }
      }
    }
  }
}